requirements.txt
# API clients
requests>=2.26.0
openai>=1.0.0
anthropic>=0.7.0

# Configuration
python-dotenv>=0.19.2

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0